// src/generated/model/updateStudentRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { UpdateStudentRequestGender } from './updateStudentRequestGender.ts';
import type { UpdateStudentRequestStatus } from './updateStudentRequestStatus.ts';

export interface UpdateStudentRequest {
  /**
     * @minLength 0
     * @maxLength 200
     */
  fullName: string;
  dateOfBirth: string;
  gender: UpdateStudentRequestGender;
  /**
     * @minLength 0
     * @maxLength 100
     */
  nationality: string;
  /**
     * @minLength 0
     * @maxLength 200
     */
  placeOfBirth?: string;
  /**
     * @minLength 0
     * @maxLength 100
     */
  religion?: string;
  /**
     * @minLength 0
     * @maxLength 100
     */
  passportNumber: string;
  passportExpiryDate?: string;
  /**
     * @minLength 0
     * @maxLength 30
     */
  telephoneNumber: string;
  /**
     * @minLength 0
     * @maxLength 200
     */
  email?: string;
  /** @minLength 1 */
  currentResidentialAddress: string;
  permanentResidentialAddress?: string;
  educationBackground?: string;
  /**
     * @minLength 0
     * @maxLength 200
     */
  fatherName?: string;
  /**
     * @minLength 0
     * @maxLength 200
     */
  motherName?: string;
  /**
     * @minLength 0
     * @maxLength 200
     */
  emergencyContactName: string;
  /**
     * @minLength 0
     * @maxLength 100
     */
  emergencyContactRelationship?: string;
  /**
     * @minLength 0
     * @maxLength 30
     */
  emergencyContactPhone: string;
  status?: UpdateStudentRequestStatus;
  remarks?: string;
}